// src/app/models/Examen.ts
export interface Examen {
    idexa: string;
    peso_ini?: number;
    peso_fin?: number;
    talla_ini?: number;
    talla_fin?: number;
    imc_ini?: number;
    imc_fin?: number;
    hematocrito_ini?: number;
    hematocrito_fin?: number;
    hemoglobina_ini?: number;
    hemoglobina_fin?: number;
    fecha_ini?: string;
    fecha_fin?: string;
    resultado_ini?: string;
    resultado_fin?: string;      
    dni_paciente?: string;    
}